requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
Pillow
streamlit-image-coordinates
gdown